[IMP] website_sale(_stock): move wishlist and stock notification
Moving anything that has to do with notifications and wishlist to website_sale.

task 5418128
