+++ addons/website_sale/controllers/main.py
@@ -946,6 +946,7 @@ class WebsiteSale(payment_portal.PaymentPortal):
             "view_track": view_track,
             "markup_data_json": json_scriptsafe.dumps(markup_data, indent=2),
             "shop_path": SHOP_PATH,
+            "user_email": request.env.user.email or request.session.get("stock_notification_email", ""),
         }
 
     def _prepare_breadcrumb_markup_data(self, base_url, category):
