+++ addons/website_sale/data/ir_cron_data.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
-    <record id="ir_cron_send_availability_email" model="ir.cron">
+    <record id="ir_cron_send_abandoned_email" model="ir.cron">
         <field name="name">eCommerce: send email to customers about their abandoned cart</field>
         <field name="interval_number">1</field>
         <field name="interval_type">hours</field>
@@ -29,4 +29,12 @@
         <field name="state">code</field>
     </record>
 
+    <record id="ir_cron_send_availability_email" model="ir.cron">
+        <field name="name">Product: send email regarding products availability</field>
+        <field name="interval_number">1</field>
+        <field name="interval_type">hours</field>
+        <field name="model_id" ref="model_product_product"/>
+        <field name="code">model._send_availability_email()</field>
+        <field name="state">code</field>
+    </record>
 </odoo>
