+++ addons/website_sale/models/__init__.py
@@ -11,6 +11,7 @@ from . import (
     payment_transaction,
     product_attribute,
     product_attribute_category,
+    product_combo,
     product_document,
     product_feed,
     product_image,
