+++ addons/website_sale/templates/product_page_templates.xml
@@ -550,6 +550,7 @@
 
     <template id="website_sale.cta_wrapper" name="CTA Wrapper">
         <div class="availability_messages o_not_editable"/>
+        <input id="wsale_user_email" type="hidden" t-att-value="user_email"/>
         <div
             id="o_wsale_cta_wrapper"
             t-att-class="'d-flex flex-wrap align-items-center gap-2 mb-4' +
