+++ addons/website_sale_collect/static/src/xml/product_availability.xml
@@ -2,7 +2,7 @@
 
 <templates>
 
-    <t t-inherit="website_sale_stock.product_availability" t-inherit-mode="extension">
+    <t t-inherit="website_sale.product_availability" t-inherit-mode="extension">
         <div id="threshold_message" position="attributes">
             <attribute name="t-elif" add="!show_click_and_collect_availability" separator=" and "/>
         </div>
