+++ addons/website_sale_stock/controllers/__init__.py
@@ -1,3 +1,3 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import location_selector, main, website_sale
+from . import location_selector
