--- addons/website_sale_stock/data/ir_cron_data.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="ir_cron_send_availability_email" model="ir.cron">
-        <field name="name">Product: send email regarding products availability</field>
-        <field name="interval_number">1</field>
-        <field name="interval_type">hours</field>
-        <field name="model_id" ref="model_product_product"/>
-        <field name="code">model._send_availability_email()</field>
-        <field name="state">code</field>
-    </record>
-</odoo>
