+++ addons/website_sale_stock/models/__init__.py
@@ -2,15 +2,9 @@
 
 from . import (
     delivery_carrier,
-    product_combo,
-    product_feed,
-    product_product,
-    product_template,
-    product_wishlist,
     res_config_settings,
     res_partner,
     sale_order,
-    sale_order_line,
     stock_picking,
     website,
 )
