--- addons/website_sale_stock/views/website_sale_stock_templates.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <template
-        id="website_sale_stock.website_sale_stock_cart_lines"
-        inherit_id="website_sale.cart_lines_quantity"
-        name="Shopping Cart Lines"
-    >
-        <xpath expr="//input[@type='text'][hasclass('quantity')]" position="attributes">
-            <attribute name="t-att-data-max">line._get_max_line_qty()</attribute>
-        </xpath>
-    </template>
-
-    <template
-        id="website_sale_stock.website_sale_stock_product"
-        inherit_id="website_sale.cta_wrapper"
-        priority="4"
-    >
-        <xpath expr="//div[@id='o_wsale_cta_wrapper']" position="before">
-            <input id="wsale_user_email" type="hidden" t-att-value="user_email"/>
-        </xpath>
-    </template>
-
-</odoo>
